refactor: 패키지 구조 정리 및 이동
- entity 및 repository 패키지가 constant 패키지 하위에 있어서 수정함
