--- src/main/kotlin/com/jiyeon/portfolio/domain/constant/entity/BaseEntity.kt
@@ -1,7 +1,0 @@
-package com.jiyeon.portfolio.domain.constant.entity
-
-import jakarta.persistence.MappedSuperclass
-
-@MappedSuperclass
-abstract class BaseEntity {
-}
