+++ src/main/kotlin/com/jiyeon/portfolio/domain/entity/Experience.kt
@@ -1,4 +1,4 @@
-package com.jiyeon.portfolio.domain.constant.entity
+package com.jiyeon.portfolio.domain.entity
 
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
@@ -6,7 +6,7 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 
 @Entity
-class ProjectDetail : BaseEntity() {
+class Experience : BaseEntity() {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
