+++ src/main/kotlin/com/jiyeon/portfolio/domain/entity/Link.kt
@@ -1,4 +1,4 @@
-package com.jiyeon.portfolio.domain.constant.entity
+package com.jiyeon.portfolio.domain.entity
 
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
@@ -6,7 +6,7 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 
 @Entity
-class ProjectSkill : BaseEntity() {
+class Link : BaseEntity() {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
