+++ src/main/kotlin/com/jiyeon/portfolio/domain/repository/AchievementRepository.kt
@@ -1,6 +1,6 @@
-package com.jiyeon.portfolio.domain.constant.repository
+package com.jiyeon.portfolio.domain.repository
 
-import com.jiyeon.portfolio.domain.constant.entity.Achievement
+import com.jiyeon.portfolio.domain.entity.Achievement
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface AchievementRepository : JpaRepository<Achievement, Long> {
