+++ src/main/kotlin/com/jiyeon/portfolio/domain/repository/HttpInterfaceRepository.kt
@@ -1,6 +1,6 @@
-package com.jiyeon.portfolio.domain.constant.repository
+package com.jiyeon.portfolio.domain.repository
 
-import com.jiyeon.portfolio.domain.constant.entity.HttpInterface
+import com.jiyeon.portfolio.domain.entity.HttpInterface
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface HttpInterfaceRepository : JpaRepository<HttpInterface, Long> {
