+++ src/main/kotlin/com/jiyeon/portfolio/domain/repository/IntroductionRepository.kt
@@ -1,6 +1,6 @@
-package com.jiyeon.portfolio.domain.constant.repository
+package com.jiyeon.portfolio.domain.repository
 
-import com.jiyeon.portfolio.domain.constant.entity.Introduction
+import com.jiyeon.portfolio.domain.entity.Introduction
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface IntroductionRepository : JpaRepository<Introduction, Long> {
