+++ src/main/kotlin/com/jiyeon/portfolio/domain/repository/LinkRepository.kt
@@ -1,6 +1,6 @@
-package com.jiyeon.portfolio.domain.constant.repository
+package com.jiyeon.portfolio.domain.repository
 
-import com.jiyeon.portfolio.domain.constant.entity.Link
+import com.jiyeon.portfolio.domain.entity.Link
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface LinkRepository : JpaRepository<Link, Long> {
