+++ src/main/kotlin/com/jiyeon/portfolio/domain/repository/ProjectSkillRepository.kt
@@ -1,6 +1,6 @@
-package com.jiyeon.portfolio.domain.constant.repository
+package com.jiyeon.portfolio.domain.repository
 
-import com.jiyeon.portfolio.domain.constant.entity.ProjectSkill
+import com.jiyeon.portfolio.domain.entity.ProjectSkill
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface ProjectSkillRepository : JpaRepository<ProjectSkill, Long> {
