+++ src/main/kotlin/com/jiyeon/portfolio/domain/repository/SkillRepository.kt
@@ -1,6 +1,6 @@
-package com.jiyeon.portfolio.domain.constant.repository
+package com.jiyeon.portfolio.domain.repository
 
-import com.jiyeon.portfolio.domain.constant.entity.Skill
+import com.jiyeon.portfolio.domain.entity.Skill
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface SkillRepository : JpaRepository<Skill, Long> {
